+++ clients/mapiclient/tomograph.c
@@ -1948,7 +1948,7 @@ main(int argc, char **argv)
 			if (*response) {
 				if (debug)
 					fprintf(stderr,"LASTLINE:%s", response);
-				len = strlen(response);
+				len = (int)strlen(response);
 				strncpy(buf, response, len + 1);
 			} else
 				len = 0;
